添加了gltfmodel的加载

+++ UniversalRenderPipeline/Sources/ShDescriptors.cpp
@@ -190,5 +190,5 @@ void ShDescriptorWriter::overwrite(VkDescriptorSet& set) {
 	for (auto& write : writes) {
 		write.dstSet = set;
 	}
-	vkUpdateDescriptorSets(pool.shDevice.device(), writes.size(), writes.data(), 0, nullptr);
+	vkUpdateDescriptorSets(pool.shDevice.device(), (uint32_t)writes.size(), writes.data(), 0, nullptr);
 }
